requirements.txt
Flask==3.1.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.6.1
urllib3==2.3.0
scikit-surprise==1.1.4
joblib>=1.2.0
pybind11>=2.12
streamlit==1.44.1
flake8
black 
pytest